test(web): create farmer service unit tests

+++ farmers-web/src/app/feature/farmer/services/farmer.service.ts
@@ -1,8 +1,8 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Observable, Subject, tap } from 'rxjs';
+import { Observable, Subject, tap, throwError } from 'rxjs';
 import { environment } from 'src/environments/environment';
-import { Farmer } from '../models/farmer.interface';
+import { Farmer, UnpersistedFarmer } from '../models/farmer.interface';
 
 const farmersEndpoint = `${environment.api}/api/farmers`;
 
@@ -19,15 +19,23 @@ export class FarmerService {
       .pipe(tap((farmerList: Farmer[]) => (this._currentFarmerList = farmerList)));
   }
 
-  createFarmer(farmer: Farmer): Observable<Farmer> {
+  createFarmer(farmer: UnpersistedFarmer): Observable<Farmer> {
     return this.httpClient.post<Farmer>(farmersEndpoint, farmer);
   }
 
   updateFarmer(farmer: Farmer): Observable<Farmer> {
+    if (farmer.id <= 0) {
+      return throwError(() => new TypeError('The given farmer object must have a valid ID'));
+    }
+
     return this.httpClient.put<Farmer>(`${farmersEndpoint}/${farmer.id}`, farmer);
   }
 
   deleteFarmer(farmerId: number): Observable<void> {
+    if (farmerId <= 0) {
+      return throwError(() => new TypeError('The given farmer ID is invalid'));
+    }
+
     return this.httpClient.delete<void>(`${farmersEndpoint}/${farmerId}`);
   }
